bootstrap: move policy setup code to internal/process/policy.js
To keep pre_execution.js focusing on bootstrap instead of
domain-specific logic.
