+++ lib/internal/bootstrap/pre_execution.js
@@ -5,7 +5,6 @@ const {
   ObjectDefineProperties,
   ObjectDefineProperty,
   ObjectGetOwnPropertyDescriptor,
-  SafeMap,
   SafeWeakMap,
   StringPrototypeStartsWith,
   globalThis,
@@ -24,9 +23,6 @@ const {
 } = require('internal/util');
 
 const { Buffer } = require('buffer');
-const {
-  ERR_MANIFEST_ASSERT_INTEGRITY,
-} = require('internal/errors').codes;
 const assert = require('internal/assert');
 
 function prepareMainThreadExecution(expandArgv1 = false,
@@ -462,51 +458,12 @@ function initializeClusterIPC() {
 }
 
 function initializePolicy() {
-  const experimentalPolicy = getOptionValue('--experimental-policy');
-  if (experimentalPolicy) {
+  const policyPath = getOptionValue('--experimental-policy');
+  if (policyPath) {
     process.emitWarning('Policies are experimental.',
                         'ExperimentalWarning');
-    const { pathToFileURL, URL } = require('internal/url');
-    // URL here as it is slightly different parsing
-    // no bare specifiers for now
-    let manifestURL;
-    if (require('path').isAbsolute(experimentalPolicy)) {
-      manifestURL = new URL(`file://${experimentalPolicy}`);
-    } else {
-      const cwdURL = pathToFileURL(process.cwd());
-      cwdURL.pathname += '/';
-      manifestURL = new URL(experimentalPolicy, cwdURL);
-    }
-    const fs = require('fs');
-    const src = fs.readFileSync(manifestURL, 'utf8');
-    const experimentalPolicyIntegrity = getOptionValue('--policy-integrity');
-    if (experimentalPolicyIntegrity) {
-      const SRI = require('internal/policy/sri');
-      const { createHash, timingSafeEqual } = require('crypto');
-      const realIntegrities = new SafeMap();
-      const integrityEntries = SRI.parse(experimentalPolicyIntegrity);
-      let foundMatch = false;
-      for (let i = 0; i < integrityEntries.length; i++) {
-        const {
-          algorithm,
-          value: expected
-        } = integrityEntries[i];
-        const hash = createHash(algorithm);
-        hash.update(src);
-        const digest = hash.digest();
-        if (digest.length === expected.length &&
-          timingSafeEqual(digest, expected)) {
-          foundMatch = true;
-          break;
-        }
-        realIntegrities.set(algorithm, digest.toString('base64'));
-      }
-      if (!foundMatch) {
-        throw new ERR_MANIFEST_ASSERT_INTEGRITY(manifestURL, realIntegrities);
-      }
-    }
-    require('internal/process/policy')
-      .setup(src, manifestURL.href);
+    const policyIntegrity = getOptionValue('--policy-integrity');
+    require('internal/process/policy').setup(policyPath, policyIntegrity);
   }
 }
